Move rest of V2Value.java to their own classes
